Refactoring
Refactored several spots where item was being spread in the return unnecessarily. 

An issue was occurring when using the typeof function. In my conditional statements I was using String or Number, when typeof results in a string containing the typeof. This was all irrelevant because when comparing a statement such as ( "1" > 0 ) this results in true.

However, now the parseInt is working and we are comparing int to int, even tho string to int worked.